+++ CHANGELOG.md
@@ -1,7 +1,14 @@
+## 2.5.0 (October 1, 2021)
+
+* Add New Select action
+* Deprecate old Select action
+
 ## 2.4.5 (September 1, 2021)
+
 Open only one connection pool per one execution and reuse it
 
 ## 2.4.4 (August 12, 2021)
+
 * Remove dependencyCheckAnalyze task
 
 ## 2.4.3 (February 12, 2021)